Refactor useServiceTheme imports to a centralized location for consistency
