+++ src/app/incore/services/[service]/ApproachSection.tsx
@@ -8,8 +8,8 @@ import InVolveApproach from "@public/inCore/approach/involveApproach.svg";
 import Cube from "@public/inCore/cube.svg";
 import CircledLine from "@public/inCore/text-circled-line.svg";
 import { useParams } from "next/navigation";
+import { useServiceTheme } from "../../../../hooks/useServiceTheme";
 import { useServiceContent } from "./hooks/useServiceContent";
-import { useServiceTheme } from "./hooks/useServiceTheme";
 
 const approachImages = {
   instack: InStackApproach,
